factor out history support, add `getSignedTx` helper
signed transactions are used as auth mechanism for different
things: tos handling, trade history and websocket auth.

in the future, probably more endpoints are going to be added.

for auth mechanism like scatter, this means right now, that a login
to the ui needs multiple transactions signed by the user, which
results in a bad ux.

to provide maximum flexibilty in the future, this removes
the built-in moonbeam request wrapper. a client can request a
tx with `getSignedTx` now, that can be used for their tos, ws-auth
and history handling, until the tx expires by the defined
timeframe.